course_service_fastapi/requirements.txt
fastapi~=0.115.2
pydantic~=2.9.2
pymongo~=4.9.2
uvicorn~=0.31.1
google-auth~=2.22.0
google-api-python-client~=2.97.0
python-multipart~=0.0.12
python-dotenv~=1.0.1
langchain-core~=0.3.12
langchain~=0.3.4
langchain_openai~=0.2.3
pika~=1.3.2
python-jose~=3.3.0
boto3~=1.35.57
motor
